+++ VersionHistory.md
@@ -6,12 +6,14 @@ Describe changes here when they're committed to the `master` branch. Move them t
 
 Prefix the description of the change with `[major]`, `[minor]` or `[patch]` in accordance with [Semantic Versioning](https://semver.org/).
 
+## Released
+
+### 0.4.0
+
 * [minor] Use `DotNetBuildOptions` to access and/or override build options.
 * [minor] Automatically run any target named `default`.
 * [minor] Return existing target; allow multiple actions per target.
 * [minor] Don't crash if a packaged project doesn't output a DLL.
-
-## Released
 
 ### 0.3.1